Make footer more responsive and not partially covered on mobile devices
There were scenarios in which social media buttons could partially cover the footer on mobile devices. This was down to the footer wrapping at the wrong screen widths and the footer not having enough of a margin for the floating social medial buttons to sit at the bottom of the page without covering the footer, see #6 for further info.

Fixes #6